Proecdure for creating booking amount
Once a customer successfully completes a booking transaction, the Booking.Start_Date and Booking.End_Date will be used to calculate the Booking_Amount. Next, currently applicable promotions/discounts will be applied to this transaction. The booking amount will be further updated to include the rental insurance charge if the customer opts to purchase that service. Finally, a % of blanket tax will be added to obtain the expected total charge, provided the customer will return the car on time.