Procedure for generating quaterly report
This procedure generates a quarterly report. It generates information regarding the number of bookings and the total revenue generated for a particular zip code in the current quarter. This procedure takes a Date parameter as an input, which can be any date in a particular quarter and the report is generated accordingly. If no input is passed the current quarter report gets generated.