Trigger to update car availability
When a car is returned by a customer, we update the status of its availability for booking as “A”, if the drop off branch is the same as branch it belongs to. Similarly, when the booking status is changed from a “Booked” state to “Active” state, indicating the customer has taken the car, we update the status of the car to “N” indicating not available for booking at this point.